+++ app/preload.js
@@ -1,14 +1,4 @@
-const { Dropdown } = require('materialize-css')
-
 // events
-document.addEventListener('DOMContentLoaded', () => {
-  const elems = document.querySelectorAll('.dropdown-trigger')
-  Dropdown.init(elems, { 
-    constrainWidth: false,
-    coverTrigger: false
-  })
-})
-
 document.addEventListener('keypress', (e) => {
   if (e.code === "Enter") {
     e.preventDefault()